Added window class

+++ espresso/utils/threading/SystemManager.h
@@ -58,7 +58,7 @@ namespace Espresso::Threading
 										   (void*)system->GetName().c_str());
 			}
 
-			cv.wait(lock, []() { return shutdowns.load() == systems.size(); });
+			cv.wait(lock, []() { return shutdowns.load() == (int)systems.size(); });
 
 			systems.clear();
 		}
